# swagger.json
{  
   "swagger":"2.0",
   "info":{  
      "description":"A golang web application, accepts an csv file, processes it and sends back a result as json",
      "version":"1.0.0",
      "title":"Billable Rates"
   },
   "host":"localhost:8100",
   "basePath":"/api",
   "schemes":[  
      "http"
   ],
   "paths":{  
      "/ping":{  
         "get":{  
            "summary":"Check connectivity",
            "description":"",
            "operationId":"ping",
            "produces":[  
               "application/json"
            ],
            "responses":{  
               "200":{  
                  "description":"Ping successful!",
                  "$ref":"#/definitions/pingResponse"
               }
            }
         }
      },
      "/invoice":{  
         "post":{  
            "summary":"Returns companies billable invoice",
            "description":"This returns an array of employee invoice",
            "operationId":"GenerateInvoice",
            "consumes":[  
               "multipart/form-data"
            ],
            "content" : "multipart/form-data",
            "produces":[  
               "application/json",
               "multipart/form-data"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"body",
                  "description":"Employees billable rate in csv",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/billableReport"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"successful operation",
                  "schema":{  
                     "$ref":"#/definitions/billableInvoice"
                  }
               }
            }
         }
      }
   },
   "definitions":{  
      "billableReport":{  
         "type":"object",
         "required":[  
            "csvReport"
         ],
         "properties":{ 
            "csvReport":{  
               "type":"string",
               "format": "byte" ,
               "example":"file"
            }
         }
      },
      "billableInvoice":{  
         "type":"object",
         "properties":{  
            "ok" : {"type":"boolean"},
            "code" : {"type" : "integer"},
            "message" : {"type" : "string"},
            "data" : {"type":"array", "items": {"type": "object", "properties":{
               "employeeId" : {"type" : "integer"},
               "workHours" : {"type" : "integer"},
               "unitPrice" : {"type" : "integer"},
               "cost" : {"type" : "integer"}
            }} }
         }
      },
      "pingResponse" : {
         "type":"object",
         "properties":{  
            "ok" : {"type":"boolean"},
            "code" : {"type" : "integer"},
            "message" : {"type" : "string"}
         }
      }
   }
}